+++ package.json
@@ -49,7 +49,6 @@
     "cross-env": "7.0.3",
     "dotenv": "10.0.0",
     "express": "4.17.1",
-    "morgan": "^1.10.0",
     "swagger-ui-express": "4.1.6",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.4",
