+++ src/resources/boards/board.router.ts
@@ -1,6 +1,6 @@
 import express, { Request, Response, Router } from 'express';
 
-import { responseWrapper } from '../../error';
+import { responseWrapper } from '../../errorHandler';
 import {
   getAllBoards,
   getBoard,
@@ -20,7 +20,6 @@ router.route('/:id').get(async (req, res) => {
   responseWrapper(res, async () => {
     const { id } = req.params;
     const board = await getBoard(id);
-    // res.status(200).send(Board.toResponse(board));
     res.status(200).send(board);
   });
 });
@@ -34,9 +33,6 @@ router.route('/:id').put(async (req, res) => {
   const { id } = req.params;
   const board = await updateBoard(id, req.body);
   res.status(200).send(board);
-
-  // const board = await createBoard(Board.fromRequest(req.body));
-  // res.status(201).send(Board.toResponse(board));
 });
 
 router.route('/:id').delete(async (req, res) => {
